requirements.txt
accelerate==0.33.0
peft==0.12.0
trl==0.9.6
transformers==4.44.0
discord.py==2.4.0
python-dotenv==1.0.1
Flask==3.0.3
datasets==2.20.0
bitsandbytes==0.43.3
pandas==2.2.2